refactored small issues with relationships, tried establishing sales_engine spec connection
Many of the method names were slightly off as well as argument status and things of the sort. Spec harness requires a Sales Engine class with an initialize argument == 1 and a startup method which builds all the repositories. We are finishing that bit up now, so after we merge, it will very simply be harnessed. Whatever happens after is up to baby jesus